+++ src/pvenv/main.py
@@ -37,6 +37,9 @@ def parse_args():
     invenv_parser.add_argument("-f", "--files", action="append", default=[])
     subparsers.add_parser("out", help="List venvs")
 
+    avenv_parser = subparsers.add_parser("activate", help="Activate a venv")
+    avenv_parser.add_argument("venv")
+    avenv_parser.add_argument("-n", "--no-cd", dest="cd", action="store_false")
     subparsers.add_parser("list", help="List venvs")
     mkvenv_parser = subparsers.add_parser("make", help="Make a new venv")
     mkvenv_parser.add_argument("venv", nargs="?", default=Path().absolute().name)
@@ -56,6 +59,8 @@ def main() -> None:
         module = subcommands.invenv
     elif args.command == "out":
         module = subcommands.outvenv
+    elif args.command == "activate":
+        module = subcommands.avenv
     elif args.command == "list":
         module = subcommands.lsvenv
     elif args.command == "make":