+++ src/pvenv/subcommands/__init__.py
@@ -1,3 +1,3 @@
-from pvenv.subcommands import invenv, lsvenv, mkvenv, outvenv, rmvenv
+from pvenv.subcommands import avenv, invenv, lsvenv, mkvenv, outvenv, rmvenv
 
-__all__ = ["invenv", "lsvenv", "mkvenv", "outvenv", "rmvenv"]
+__all__ = ["avenv", "invenv", "lsvenv", "mkvenv", "outvenv", "rmvenv"]